Removes non-compact rank symbols in favor of compact rank
