+++ CHANGELOG
@@ -1,4 +1,5 @@
 [master]
+    - Removes non-compact rank symbols in favor of compact rank
     - All duplicate whitespaces / tabs / newlines are now removed from non-multiline fields
     - Fixed a bug in the IEEEXploreFetcher
     - Replaced some deprecated DefaultFormBuilder with FormBuilder
