+++ src/main/java/net/sf/jabref/gui/GUIGlobals.java
@@ -155,7 +155,6 @@ public class GUIGlobals {
 
     public static final int DEFAULT_FIELD_LENGTH = 100;
     public static final int NUMBER_COL_LENGTH = 32;
-    public static final int WIDTH_ICON_COL_RANKING = 35; // Width of Ranking Icon Column
     public static final int WIDTH_ICON_COL = 19;
 
     // Column widths for export customization dialog table:
