+++ src/main/java/net/sf/jabref/specialfields/Rank.java
@@ -16,20 +16,41 @@
 package net.sf.jabref.specialfields;
 
 import net.sf.jabref.Globals;
+import net.sf.jabref.gui.IconTheme;
 import net.sf.jabref.logic.l10n.Localization;
 
-public abstract class Rank extends SpecialField {
+import javax.swing.*;
+import java.util.ArrayList;
 
-    Rank() {
+public class Rank extends SpecialField {
+
+    private static Rank INSTANCE;
+
+    private Rank() {
         TEXT_DONE_PATTERN = "Set rank to '%0' for %1 entries";
+
+        ArrayList<SpecialFieldValue> values = new ArrayList<SpecialFieldValue>();
+        //lab.setName("i");
+        values.add(new SpecialFieldValue(this, null, "clearRank", Localization.lang("Clear rank"), null, Localization.lang("No rank information")));
+        // DO NOT TRANSLATE "rank1" etc. as this makes the .bib files non portable
+        values.add(new SpecialFieldValue(this, "rank1", "setRank1", Localization.lang("Set rank to one star"), IconTheme.getImage("rank1"), Localization.lang("One star")));
+        values.add(new SpecialFieldValue(this, "rank2", "setRank2", Localization.lang("Set rank to two stars"), IconTheme.getImage("rank2"), Localization.lang("Two stars")));
+        values.add(new SpecialFieldValue(this, "rank3", "setRank3", Localization.lang("Set rank to three stars"), IconTheme.getImage("rank3"), Localization.lang("Three stars")));
+        values.add(new SpecialFieldValue(this, "rank4", "setRank4", Localization.lang("Set rank to four stars"), IconTheme.getImage("rank4"), Localization.lang("Four stars")));
+        values.add(new SpecialFieldValue(this, "rank5", "setRank5", Localization.lang("Set rank to five stars"), IconTheme.getImage("rank5"), Localization.lang("Five stars")));
+        this.setValues(values);
     }
 
     public static Rank getInstance() {
-        if (Globals.prefs.getBoolean(SpecialFieldsUtils.PREF_RANKING_COMPACT)) {
-            return RankCompact.getInstance();
-        } else {
-            return RankExtended.getInstance();
+        if (Rank.INSTANCE == null) {
+            Rank.INSTANCE = new Rank();
         }
+        return Rank.INSTANCE;
+    }
+
+    @Override
+    public Icon getRepresentingIcon() {
+        return IconTheme.getImage("ranking");
     }
 
     @Override
