+++ src/main/resources/images/crystal_16/Icons.properties
@@ -72,11 +72,6 @@ rank2=rank2.png
 rank3=rank3.png
 rank4=rank4.png
 rank5=rank5.png
-rankc1=rankc1.png
-rankc2=rankc2.png
-rankc3=rankc3.png
-rankc4=rankc4.png
-rankc5=rankc5.png
 ranking=ranking.png
 readStatusRead=read.png
 readStatusSkimmed=readskimmed.png
