+++ src/main/resources/l10n/JabRef_de.properties
@@ -2259,7 +2259,6 @@ Bibtex_key_patterns=BibTeX-Key-Muster
 Changed_special_field_settings=Einstellungen_f\u00fcr_spezielle_Felder_ge\u00e4ndert
 Clear_priority=Priorit\u00e4t_aufheben
 Clear_rank=Rang_aufheben
-Compact_rank=Kompakter_Rang
 Convert_1st,_2nd,_..._to_real_superscripts=1st,_2nd,_..._in_hochgestellte_Buchstaben_konvertieren
 Dropped_comment_from_database=Kommentar_aus_Datei_gel\u00f6scht
 Enable_PDF_preview=PDF-Vorschau_aktivieren
