+++ src/main/resources/l10n/JabRef_en.properties
@@ -2246,7 +2246,6 @@ Bibtex_key_patterns=Bibtex_key_patterns
 Changed_special_field_settings=Changed_special_field_settings
 Clear_priority=Clear_priority
 Clear_rank=Clear_rank
-Compact_rank=Compact_rank
 Convert_1st,_2nd,_..._to_real_superscripts=Convert_1st,_2nd,_..._to_real_superscripts
 Dropped_comment_from_database=Dropped_comment_from_database
 Enable_PDF_preview=Enable_PDF_preview
