+++ src/main/resources/l10n/JabRef_es.properties
@@ -1403,7 +1403,6 @@ Bibtex_key_patterns=Patrones_de_clave_BibTeX
 Changed_special_field_settings=Ajustes_de_campos_especiales_cambiados
 Clear_priority=Limpiar_prioridad
 Clear_rank=Limpiar_rango
-Compact_rank=Compactar_rango
 Convert_1st,_2nd,_..._to_real_superscripts=Convertir_1st,_2nd,_..._a_super\u00edndices
 Dropped_comment_from_database=Eliminado_comentario_de_la_base_de_datos
 Enable_PDF_preview=Habilitar_vista_previa_de_PDF
