+++ src/main/resources/l10n/JabRef_fr.properties
@@ -1442,7 +1442,6 @@ Bibtex_key_patterns=Mod\u00e8les_de_clef_BibTeX
 Changed_special_field_settings=Param\u00e8tres_de_champs_sp\u00e9ciaux_modifi\u00e9s
 Clear_priority=Effacer_la_priorit\u00e9
 Clear_rank=Effacer_le_rang
-Compact_rank=Compacter_le_rang
 Convert_1st,_2nd,_..._to_real_superscripts=Convertir_1er,__2nd,_..._en_vrais_exposants
 Dropped_comment_from_database=Commentaires_supprim\u00e9s_de_la_base
 Enable_PDF_preview=Autoriser_la_pr\u00e9visualisation_des_PDF
