+++ src/main/resources/l10n/JabRef_it.properties
@@ -1524,7 +1524,6 @@ Changed_special_field_settings=Cambiate_le_impostazioni_dei_campi_speciali
 Clear_priority=Azzera_le_priorit\u00e0
 # not sure how to translate 'rank': voto, punteggio, valutazione, grado, rango?
 Clear_rank=Azzera_la_valutazione
-Compact_rank=Compatta_la_valutazione
 Convert_1st,_2nd,_..._to_real_superscripts=Converti_1st,_2nd,_..._in_apici
 Dropped_comment_from_database=Cancellato_il_commento_dal_database
 Enable_PDF_preview=Abilita_anteprima_PDF
