+++ src/main/resources/l10n/JabRef_ja.properties
@@ -2245,7 +2245,6 @@ Bibtex_key_patterns=BibTeX\u9375\u30d1\u30bf\u30fc\u30f3
 Changed_special_field_settings=\u7279\u6b8a\u30d5\u30a3\u30fc\u30eb\u30c9\u306e\u8a2d\u5b9a\u304c\u5909\u66f4\u3055\u308c\u307e\u3057\u305f
 Clear_priority=\u512a\u5148\u5ea6\u3092\u6d88\u53bb
 Clear_rank=\u8a55\u4fa1\u3092\u6d88\u53bb
-Compact_rank=\u8a55\u4fa1\u3092\u7c21\u7d20\u8868\u793a
 Convert_1st,_2nd,_..._to_real_superscripts=1st,_2nd,_..._\u7b49\u3092\u4e0b\u4ed8\u304d\u6587\u5b57\u3092\u4f7f\u3046\u3088\u3046\u5909\u63db
 Dropped_comment_from_database=\u30c7\u30fc\u30bf\u30d9\u30fc\u30b9\u304b\u3089\u30b3\u30e1\u30f3\u30c8\u3092\u53d6\u308a\u9664\u304d\u307e\u3057\u305f
 Enable_PDF_preview=PDF\u30d7\u30ec\u30d3\u30e5\u30fc\u3092\u6709\u52b9\u306b\u3059\u308b
