+++ src/main/resources/l10n/JabRef_pt_BR.properties
@@ -1410,7 +1410,6 @@ Bibtex_key_patterns=Padr\u00f5es_de_chaves_BibTeX
 Changed_special_field_settings=Prefer\u00eancias_de_campos_especiais_alteradas
 Clear_priority=Limpar_prioridades
 Clear_rank=Limpar_classifica\u00e7\u00e3o
-Compact_rank=Compactar_classifica\u00e7\u00e3o
 Convert_1st,_2nd,_..._to_real_superscripts=Converter_primeiro_segundo_..._para_sobrescritos
 Dropped_comment_from_database=Coment\u00e1rio_retirado_da_base_de_dados
 Enable_PDF_preview=Habilitar_preview_de_PDF
