+++ src/main/resources/l10n/JabRef_ru.properties
@@ -2244,7 +2244,6 @@ Bibtex_key_patterns=Шаблоны_ключа_Bibtex
 Changed_special_field_settings=Настройки_полей_особых_отметок_изменены
 Clear_priority=Очистить_значения_приоритета
 Clear_rank=Очистить_значения_ранга
-Compact_rank=Свернуть_значения_ранга
 Convert_1st,_2nd,_..._to_real_superscripts=Преобразовать_1-й,_2-й,_..._в_надстрочные_знаки
 Dropped_comment_from_database=Комментари,_отброшенные_БД
 Enable_PDF_preview=Разрешить_предпросмотр_PDF
