+++ src/main/resources/l10n/JabRef_tr.properties
@@ -1433,7 +1433,6 @@ Bibtex_key_patterns=Bibtex_anahtar_paternleri
 Changed_special_field_settings=De\u011fi\u015ftirilmi\u015f_\u00f6zel_alan_ayarlar\u0131
 Clear_priority=\u00d6nceli\u011fi_sil
 Clear_rank=R\u00fctbeyi_sil
-Compact_rank=R\u00fctbeyi_s\u0131k\u0131la\u015ft\u0131r
 Convert_1st,_2nd,_..._to_real_superscripts=1.,_2.,_..._vs.'yi_ger\u00e7ek_\u00fcst_karaktere_d\u00f6n\u00fc\u015ft\u00fcr
 Dropped_comment_from_database=Yorum_veri_taban\u0131ndan_\u00e7\u0131kart\u0131ld\u0131
 Enable_PDF_preview=PDF_\u00f6nizlemeyi_etkinle\u015ftir
