+++ src/main/resources/l10n/JabRef_vi.properties
@@ -2241,7 +2241,6 @@ Bibtex_key_patterns=
 Changed_special_field_settings=
 Clear_priority=
 Clear_rank=
-Compact_rank=
 Convert_1st,_2nd,_..._to_real_superscripts=
 Dropped_comment_from_database=
 Enable_PDF_preview=
